pyproject.toml
[project]
name = "ex8-producer"
version = "0.1.0"
description = ""
authors = [
    {name = "Matheus Perira",email = "pnmatheus01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3",
    "python-dotenv",
    "avro",
    "kafka-python",
]

[tool.poetry]
packages = [{include = "ex8_producer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
